+++ spec/lib/type/array_spec.rb
@@ -6,10 +6,12 @@ RSpec.describe Scheming::Type::Array do
 
   it do
     expect(instance.type).to eq(type)
+    expect(Scheming::Schema.json(instance))
+      .to be_valid_json_schema
     expect(Scheming::Schema.json(instance)).to eq(
       {
         type: 'array',
-        items: { type: 'numeric' }
+        items: { type: 'number' }
       }
     )
   end