+++ spec/lib/type/enum_spec.rb
@@ -8,8 +8,9 @@ RSpec.describe Scheming::Type::Enum do
   it do
     expect(instance.type).to be(type)
     expect(instance.values).to be(values)
+    expect(Scheming::Schema.json(instance)).to be_valid_json_schema
     expect(Scheming::Schema.json(instance)).to eq(
-      enum: values,
+      enum: values.to_a,
       type: 'string'
     )
   end