+++ spec/lib/type/integer_spec.rb
@@ -5,6 +5,7 @@ RSpec.describe Scheming::Type::Integer do
 
   it do
     expect(instance).to be_a(described_class)
+    expect(Scheming::Schema.json(instance)).to be_valid_json_schema
     expect(Scheming::Schema.json(instance)).to eq(type: 'integer')
   end
 end